+++ backend/src/main/java/com/smartqueue/backend/controller/WebSocketMessageController.java
@@ -17,7 +17,7 @@ public class WebSocketMessageController {
 
     @MessageMapping("/queue.subscribe/{officeId}")
     public void onSubscribe(@DestinationVariable Integer officeId) {
-        QueueStateDTO state = queueService.getQueueState(officeId);
+        QueueStateDTO state = queueService.getQueueState((long)officeId);
         broadcastService.broadcastQueueState(officeId, state);
     }
 }
