+++ backend/src/main/java/com/smartqueue/backend/service/AIService.java
@@ -70,7 +70,7 @@ public class AIService {
     public ChatResponse processMessage(ChatRequest request) {
         int officeId = request.getOfficeId() != null ? request.getOfficeId() : 1;
 
-        var queueState = queueService.getQueueState(officeId);
+        var queueState = queueService.getQueueState((long)officeId);
         String context = String.format(
                 "Queue context: %d people waiting, ~%d min estimated wait. ",
                 queueState.getWaitingCount(),
@@ -220,6 +220,6 @@ public class AIService {
                 .build();
     }
     public com.smartqueue.backend.dto.QueueStateDTO getQueueStateForTelegram(int officeId) {
-        return queueService.getQueueState(officeId);
+        return queueService.getQueueState((long)officeId);
     }
 }
